v1.40.23 Released!
- Added new exchange Bittrex
- Portable mode for Linux and macOS
- Secure auto update for Linux x86_64
- Uninstall option in help menu
- Fixed Binance own trade history order
- Minor fixes

+++ src/historyitem.h
@@ -40,7 +40,7 @@ public:
     HistoryItem();
 
     bool displayFullDate;
-    quint32 dateTimeInt;
+    qint64  dateTimeInt;
     quint32 dateInt;
     QString dateTimeStr;
     QString timeStr;
